% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{apis}
\alias{apis}
\alias{get_api_url}
\title{AI APIs}
\usage{
get_api_url(provider, name = NULL)
}
\arguments{
\item{provider}{A character string specifying the API provider.}

\item{name}{A character string specifying the service name.}
}
\value{
A character string of the base URL for the specified API service.
}
\description{
These functions are helpers for working with various AI APIs.
\itemize{
\item \code{get_api_url()}: Retrieve the base URL for a given API provider and service.
}
}
