% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_config_template}
\alias{create_config_template}
\title{Create a Configuration Template}
\usage{
create_config_template(
  config_file = Sys.getenv("R_CONFIG_FILE", unset = pkg_sys_config("config.yml")),
  template_file = NULL
)
}
\arguments{
\item{config_file}{A character string representing the path to the configuration file. If left
\code{NULL}, the function will attempt to use the default configuration file via \code{pkg_sys_config("config.yml")}.}

\item{template_file}{A character string representing the path to the template file. If left \code{NULL}, the function will
create a new file with the same name and path as the original file but with the extension \code{.template.yml}.}
}
\value{
The function writes the template configuration file to the specified location.
}
\description{
Create a template configuration file based on an existing configuration file with all values replaced by placeholders.
}
