% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid.R
\name{create_mermaid_diagram}
\alias{create_mermaid_diagram}
\title{Generate Mermaid Diagrams}
\usage{
create_mermaid_diagram(chat_agent = NULL, prompt = NULL, ...)
}
\arguments{
\item{chat_agent}{An optional chat agent object. If not provided, a default agent will be created.}

\item{prompt}{A prompt to generate the diagram. If not provided, a default prompt will be used.}

\item{...}{Additional arguments passed to \code{\link[=prompt_mermaid_user]{prompt_mermaid_user()}}. (e.g. \code{context}, \code{code}, \code{chart_type}, \code{chart_styles}).}
}
\value{
A character string containing the mermaid code block.
}
\description{
Generate a mermaid diagram using the dedicated \code{\link[=mermaid_agent]{mermaid_agent()}} based on a prompt or code.
}
\examples{
\dontrun{
# generate a diagram based off of code
example <- "
  starwars |>
    group_by(species) |>
    summarise(
      n = n(),
      mass = mean(mass, na.rm = TRUE)
    ) |>
    filter(
      n > 1,
      mass > 50
    )
"
create_mermaid_diagram(code = example)

# generate a diagram based off of a prompt
prompt <- prompt_mermaid_user("Create a flowchart representing the process of photosynthesis.")
create_mermaid_diagram(prompt = prompt)

# specify chart type and styles
create_mermaid_diagram(prompt = prompt, chart_type = "flowchart", chart_styles = "Make the chart colorful.")
}
}
