% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect}
\alias{db_connect}
\title{Connect to the Database}
\usage{
db_connect(db_config = get_db_config(), pool = TRUE, ...)
}
\arguments{
\item{db_config}{A list containing the database configuration. Must include the following fields:
\code{dbname}, \code{host}, \code{user}, \code{password}, \code{port}.}

\item{pool}{Logical. If \code{TRUE}, the connection will be pooled using \code{\link[pool:dbPool]{pool::dbPool()}}, otherwise will
create a single connection via \code{\link[DBI:dbConnect]{DBI::dbConnect()}}. Default is \code{TRUE}.}

\item{...}{Additional arguments passed to the connection function, i.e. either \code{\link[pool:dbPool]{pool::dbPool()}} or \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\value{
Database connection object.
If \code{pool = TRUE}, returns a \link[pool:Pool-class]{pool::Pool} (R6) object.
If \code{pool = FALSE}, returns a \link[DBI:DBIConnection-class]{DBI::DBIConnection} (S4) object.
}
\description{
Connect to the database using the provided configuration.
}
