% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_get_documents}
\alias{db_get_documents}
\title{Retrieve Documents}
\usage{
db_get_documents(pool, query, max_results = 10L, use_cache = TRUE)
}
\arguments{
\item{pool}{Database connection pool object.}

\item{query}{The query to search for in the documents.}

\item{max_results}{Maximum number of documents to retrieve.}

\item{use_cache}{Logical. If \code{TRUE}, will attempt to retrieve documents from the cache first.}
}
\value{
A data frame containing the retrieved documents.
}
\description{
Retrieve documents from the database based on the provided \code{query}.
}
\examples{
\dontrun{
pool <- db_connect()
db_get_documents(pool, "Anthropic Research", 5)
}
}
