% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_preload_documents}
\alias{db_preload_documents}
\title{Pre-Load Documents}
\usage{
db_preload_documents(pool, topic, max_docs = 10)
}
\arguments{
\item{pool}{Database connection pool object.}

\item{topic}{The topic to find documents for.}

\item{max_docs}{Maximum number of documents to load.}
}
\value{
A formatted string of documents
}
\description{
This function pre-loads documents for a "Cache Augmented Generation" (CAG) system, based on provided \code{topic}.
}
