% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{docker}
\alias{docker}
\title{Run \code{docker} Command}
\usage{
docker(cmd, ..., .timeout = 60, .echo = TRUE)
}
\arguments{
\item{cmd}{The command to run.}

\item{...}{Additional arguments to pass to the command.}

\item{.timeout}{The timeout in seconds for the command to complete. Default is 60 seconds.}

\item{.echo}{Logical. If \code{TRUE}, the output will be printed to the console. Default is \code{TRUE}.}
}
\value{
The output of the command.
}
\description{
This function runs a command using \code{docker} and returns the output.
}
\examples{
\dontrun{
# get docker help
docker("--help")
# list docker images
docker("images")
# run a docker container
docker("run", "--rm", "hello-world")
}
}
