% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed_openai}
\alias{embed_openai}
\title{Create an OpenAI Embedding}
\usage{
embed_openai(
  text,
  model = "text-embedding-3-small",
  api_key = get_openai_api_key()
)
}
\arguments{
\item{text}{Character string to embed.}

\item{model}{The model to use for the embedding. Default is \code{text-embedding-3-small}.}

\item{api_key}{The OpenAI API key to use for the embedding. Default is \code{get_openai_api_key()}.}
}
\value{
A numeric vector of the embedding.
}
\description{
Call the OpenAI Embeddings API to create an embedding for a given text.
}
\examples{
\dontrun{
embed_openai("Hello, world!")
}
}
