% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_base64}
\alias{encode_base64}
\alias{decode_base64}
\title{Base64 Encoding & Decoding}
\usage{
encode_base64(data)

decode_base64(data)
}
\arguments{
\item{data}{A character vector to encode or decode.}
}
\value{
\itemize{
\item \code{encode_base64()}: A character vector of Base64 encoded data.
\item \code{decode_base64()}: A character vector of decoded data.
}
}
\description{
Functions for encoding and decoding data using Base64 encoding.
}
