% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_r_code}
\alias{execute_r_code}
\title{Execute R Code}
\usage{
execute_r_code(code)
}
\arguments{
\item{code}{A character string containing the R code to execute.}
}
\value{
The result of the executed R code.
}
\description{
This function executes a given R code string and returns the result.
}
\examples{
\dontrun{
execute_r_code("print('Hello, World!')")
}

}
