% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_code}
\alias{extract_code}
\title{Extract Code}
\usage{
extract_code(text, lang = NULL, print = FALSE)
}
\arguments{
\item{text}{A character string representing the text to extract code blocks from.}

\item{lang}{The language of the code block to extract. If left \code{NULL}, the pattern will attempt to
match any language.}

\item{print}{Logical. If \code{TRUE}, the extracted code block content will be printed to the console.
Defaults to \code{FALSE}.}
}
\value{
A character string representing the extracted code block content.
}
\description{
Extract code blocks from a markdown string.

This function extracts code blocks from a markdown string using a regular expression pattern.
}
\examples{
examples/ex_extract_code.R
}
\seealso{
\code{\link[=get_regex_code_pattern]{get_regex_code_pattern()}} for more information on the regular expression pattern used to extract code blocks.
}
