% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firecrawl.R
\name{firecrawl}
\alias{firecrawl}
\alias{firecrawl_request}
\alias{firecrawl_scrape_url}
\alias{firecrawl_crawl_url}
\title{FireCrawl API}
\usage{
firecrawl_request(endpoint = NULL, api_key = get_api_key("firecrawl"))

firecrawl_scrape_url(url, api_key = get_api_key("firecrawl"))

firecrawl_crawl_url(url, ..., api_key = get_api_key("firecrawl"))
}
\arguments{
\item{endpoint}{The API endpoint to use. Default is \code{NULL}.}

\item{api_key}{The API key to use. Defaults to the value returned by \code{get_api_key("firecrawl")}.}

\item{url}{The URL to scrape or crawl.}

\item{...}{Additional arguments to pass to the API.}
}
\value{
\itemize{
\item \code{firecrawl_request()}: A request object for the FireCrawl API.
\item \code{firecrawl_scrape_url()}: The scraped content in markdown format.
\item \code{firecrawl_crawl_url()}: The response from the FireCrawl API.
}
}
\description{
The FireCrawl API provides a way to scrape and crawl web pages.
\itemize{
\item \code{firecrawl_request()}: Create a request object for the FireCrawl API.
\item \code{firecrawl_scrape_url()}: Scrape a URL using the FireCrawl API.
\item \code{firecrawl_crawl_url()}: Crawl a URL using the FireCrawl API.
}
}
