% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{format_code}
\alias{format_code}
\title{Format Code Block}
\usage{
format_code(code, keep_comments = TRUE, style = list())
}
\arguments{
\item{code}{The code block to format.}

\item{keep_comments}{Whether to keep comments (default: TRUE).}

\item{style}{A list of formatting options to override defaults.}
}
\value{
A formatted code block.
}
\description{
Uses formatR to format a code block with consistent styling.
}
