% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{geocode}
\alias{geocode}
\title{Geocode}
\usage{
geocode(location, api_key = get_api_key("openweather"))
}
\arguments{
\item{location}{A character string representing the location to geocode.}

\item{api_key}{A character string representing the OpenWeather API key.}
}
\value{
A list containing the geocoded location data.
}
\description{
Geocode a location using the OpenWeather API.
}
