% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{get_weather}
\alias{get_weather}
\title{Get Weather}
\usage{
get_weather(
  lat,
  lon,
  units = c("standard", "metric", "imperial"),
  api_key = get_api_key("openweather")
)
}
\arguments{
\item{api_key}{A character string representing the OpenWeather API key.
If not provided, it will be retrieved from the environment variable \code{OPENWEATHER_API_KEY}.}

\item{location}{A character string representing the location to get the weather for.}
}
\value{
A list containing the current weather data for the specified location.
}
\description{
Get the current weather for a given location using the OpenWeather API.
}
