% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{git}
\alias{git}
\title{Run \code{git} Command}
\usage{
git(cmd, ..., .timeout = 60, .echo = TRUE)
}
\arguments{
\item{cmd}{The command to run.}

\item{...}{Additional arguments to pass to the command.}

\item{.timeout}{The timeout in seconds for the command to complete. Default is 60 seconds.}

\item{.echo}{Logical. If \code{TRUE}, the output will be printed to the console. Default is \code{TRUE}.}
}
\value{
The output of the command.
}
\description{
This function runs a command using \code{git} and returns the output.
}
\examples{
\dontrun{
# get git version
git("--version")
# git status
git("status")
# git commit
git("commit", "-m", "Initial commit")
}
}
