% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmaps.R
\name{gmaps_extract_place_info}
\alias{gmaps_extract_place_info}
\title{Extract Place Information}
\usage{
gmaps_extract_place_info(place)
}
\arguments{
\item{place}{A list representing a place object returned by the Google Maps Places API.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{display_name}: The display name of the place.
\item \code{formatted_address}: The formatted address of the place.
\item \code{phone}: The phone number of the place.
\item \code{business_status}: The business status of the place.
\item \code{business_primary_type}: The primary type of the business.
\item \code{business_types}: A vector of business types associated with the place.
\item \code{latitude}: The latitude of the place.
\item \code{longitude}: The longitude of the place.
\item \code{maps_url}: The Google Maps URL of the place.
\item \code{place_id}: The unique identifier of the place.
\item \code{website}: The website URL of the place.
\item \code{domain}: The domain extracted from the website URL.
}
}
\description{
This function extracts relevant information from a place object returned by the Google Maps Places API.
}
