% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmaps.R
\name{gmaps_find_best_match}
\alias{gmaps_find_best_match}
\title{Determine the Best Match from Multiple Places}
\usage{
gmaps_find_best_match(
  places,
  company_name,
  company_address,
  company_phone = NULL
)
}
\arguments{
\item{places}{List of places returned from the Google Maps Places API.}

\item{company_name}{Character string representing the name of the company.}

\item{company_address}{Character string representing the address of the company.}

\item{company_phone}{Character string representing the phone number of the company.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{display_name}: The display name of the place.
\item \code{formatted_address}: The formatted address of the place.
\item \code{phone}: The phone number of the place.
\item \code{business_status}: The business status of the place.
\item \code{business_primary_type}: The primary type of the business.
\item \code{business_types}: A vector of business types associated with the place.
\item \code{latitude}: The latitude of the place.
\item \code{longitude}: The longitude of the place.
\item \code{maps_url}: The Google Maps URL of the place.
\item \code{place_id}: The unique identifier of the place.
\item \code{website}: The website URL of the place.
\item \code{domain}: The domain extracted from the website URL.
}
}
\description{
This function takes a list of places returned from the Google Maps Places API
and compares them to find the best match based on various criteria.
}
\details{
The function evaluates each place based on:
\itemize{
\item Type of place (e.g., real estate agency, property management company)
\item String similarity of the company name
\item String similarity of the address
\item Phone number area code match
\item Presence of a website
\item Business status (operational or not)
}

The function returns the place with the highest score, indicating the best match.
}
