% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmaps.R
\name{gmaps_geocode_address}
\alias{gmaps_geocode_address}
\title{Geocode Address}
\usage{
gmaps_geocode_address(address, api_key = get_gmaps_api_key())
}
\arguments{
\item{address}{A character string representing the address to geocode.}

\item{api_key}{A character string representing the Google Maps API key.
Will default to the result from \code{\link[=get_gmaps_api_key]{get_gmaps_api_key()}}.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{status}: A character string indicating the status of the API request.
\item \code{formatted_address}: A character string representing the formatted address.
\item \code{place_id}: A character string representing the place ID.
\item \code{place_types}: A character string representing the place types.
\item \code{latitude}: A numeric value representing the latitude.
\item \code{longitude}: A numeric value representing the longitude.
}
}
\description{
Geocode an address using the Google Maps Geocoding API.
}
\examples{
examples/ex_gmaps_geocode_address.R

}
