% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmaps.R
\name{gmaps_places_search}
\alias{gmaps_places_search}
\title{Google Maps Places Search}
\usage{
gmaps_places_search(
  company_name,
  company_address,
  company_phone = NULL,
  api_key = get_gmaps_api_key()
)
}
\arguments{
\item{company_name}{Character string representing the name of the company.}

\item{company_address}{Character string representing the address of the company.}

\item{company_phone}{Character string representing the phone number of the company.}

\item{api_key}{A character string representing the Google Maps API key.
Defaults to the result from \code{\link[=get_gmaps_api_key]{get_gmaps_api_key()}}.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{website}: A character string representing the website URL.
\item \code{domain}: A character string representing the domain extracted from the website URL.
\item \code{display_name}: A character string representing the display name of the place.
\item \code{formatted_address}: A character string representing the formatted address of the place.
\item \code{phone}: A character string representing the phone number of the place.
\item \code{business_status}: A character string representing the business status of the place.
\item \code{business_type}: A character string representing the business type of the place.
\item \code{latitude}: A numeric value representing the latitude of the place.
\item \code{longitude}: A numeric value representing the longitude of the place.
\item \code{maps_url}: A character string representing the Google Maps URL of the place.
\item \code{place_id}: A character string representing the place ID of the place.
}
}
\description{
Search for a company using the Google Maps Places API.
}
\examples{
\dontrun{
  gmaps_places_search("Google", "1600 Amphitheatre Parkway, Mountain View, CA")
}
}
