% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langfuse.R
\name{lf_create_generation_event}
\alias{lf_create_generation_event}
\title{Create a Generation Event}
\usage{
lf_create_generation_event(
  trace_id,
  model,
  prompt,
  completion = NULL,
  metadata = NULL
)
}
\arguments{
\item{trace_id}{Parent trace ID}

\item{model}{Model name}

\item{prompt}{Prompt text}

\item{completion}{Completion text}

\item{metadata}{Additional metadata as a list}
}
\value{
A list representing a generation event
}
\description{
Creates a generation event for Langfuse.
}
\keyword{internal}
