% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langfuse.R
\name{lf_create_score_event}
\alias{lf_create_score_event}
\title{Create a Score Event}
\usage{
lf_create_score_event(trace_id, value, comment = NULL)
}
\arguments{
\item{trace_id}{Parent trace ID}

\item{value}{Score value (0-1)}

\item{comment}{Optional comment}
}
\value{
A list representing a score event
}
\description{
Creates a score event for Langfuse.
}
\keyword{internal}
