% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langfuse.R
\name{lf_create_trace_event}
\alias{lf_create_trace_event}
\title{Create a Trace Event}
\usage{
lf_create_trace_event(
  name,
  user_id = NULL,
  input = NULL,
  output = NULL,
  metadata = NULL
)
}
\arguments{
\item{name}{Name of the trace}

\item{user_id}{User ID}

\item{input}{Input text}

\item{output}{Output text}

\item{metadata}{Additional metadata as a list}
}
\value{
A list representing a trace event
}
\description{
Creates a trace event for Langfuse.
}
\keyword{internal}
