% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_function_call}
\alias{log_function_call}
\title{Log a Function Call with Arguments}
\usage{
log_function_call(
  fn_name,
  args = list(),
  level = "INFO",
  namespace = .logger_env$logger_ns
)
}
\arguments{
\item{fn_name}{Name of the function being called}

\item{args}{List of arguments}

\item{level}{Logging level (default: "INFO")}

\item{namespace}{Logger namespace}
}
\description{
Log a Function Call with Arguments
}
