% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger_init}
\alias{logger_init}
\title{Initialize Logger}
\usage{
logger_init(
  level = logger::INFO,
  formatter = NULL,
  layout = NULL,
  log_file = NULL,
  console = TRUE,
  async = FALSE,
  namespace = NULL,
  frame = parent.frame(),
  ...
)
}
\arguments{
\item{level}{The logging level. Default is \code{logger::INFO}.}

\item{formatter}{The formatter function. Default is \code{logger::formatter_json}.}

\item{layout}{The layout function. Default is \code{logger::layout_json}.}

\item{log_file}{The log file path. Default is \code{NULL}, which uses the default log path.}

\item{console}{Whether to log to the console. Default is \code{TRUE}.}

\item{async}{Whether to use asynchronous logging. Default is \code{FALSE}.}

\item{namespace}{The namespace for the logger. Default is \code{NULL}, which uses the package name.}

\item{frame}{The parent frame for the logger. Default is \code{parent.frame()}.}

\item{...}{Additional arguments to pass to the logger.}
}
\value{
Invisibly returns a list of logger metadata variables.
}
\description{
Initializes the logger with the specified parameters.
}
