% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger_reset}
\alias{logger_reset}
\title{Reset Logger Configuration}
\usage{
logger_reset(namespace = NULL)
}
\arguments{
\item{namespace}{The namespace for the logger. Default is \code{NULL}, which uses the package name.}
}
\description{
Resets the logger configuration to the default settings and removes all hooks.
}
