% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noclocksai-package.R
\docType{package}
\name{noclocksai-package}
\alias{noclocksai}
\alias{noclocksai-package}
\title{noclocksai: No Clocks AI}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The `noclocksai` package provides a set of tools for working with AI at No Clocks, LLC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/noclocks/noclocksai}
  \item Report bugs at \url{https://github.com/noclocks/noclocksai/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Patrick Howard \email{patrick.howard@noclocks.dev} [reviewer]
}

Other contributors:
\itemize{
  \item No Clocks, LLC \email{team@noclocks.dev} [copyright holder, funder]
}

}
\keyword{internal}
