% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npx.R
\name{npx}
\alias{npx}
\title{Run \code{npx} Command}
\usage{
npx(cmd, ...)
}
\arguments{
\item{cmd}{The command to run.}

\item{...}{Additional arguments to pass to the command.}
}
\value{
The output of the command.
}
\description{
This function runs a command using \code{npx} and returns the output.
}
\examples{
\dontrun{
# get npx help
npx("--help")

# create a new react app
npx("create-react-app", "my-app")

# run repomix tool
npx("repomix", "--help")
}
}
