% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather.R
\name{openweather}
\alias{openweather}
\alias{ow_geocode}
\alias{ow_get_weather}
\title{Open Weather API}
\usage{
ow_geocode(location, api_key = get_api_key("openweather"))

ow_get_weather(
  lat,
  lon,
  units = c("standard", "metric", "imperial"),
  api_key = get_api_key("openweather")
)
}
\arguments{
\item{location}{A character string representing the location to get the weather for.}

\item{api_key}{A character string representing the Open Weather API key.
Defaults to the value returned by \code{get_api_key("openweather")}.}
}
\value{
\itemize{
\item \code{ow_geocode()}: A list containing the geocoded location data.
\item \code{ow_get_weather()}: A list containing the current weather data for the specified location.
}
}
\description{
The \href{https://openweathermap.org/api}{OpenWeather API} provides weather data for any location.

The following functions are available:
\itemize{
\item \code{ow_geocode()}: Geocode a location using the OpenWeather API.
\item \code{ow_get_weather()}: Get the current weather for a given location using the OpenWeather API.
}
}
\examples{
examples/ex_openweather.R
}
