% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{prompts}
\alias{prompts}
\alias{prompt_default_sys}
\title{Prompts}
\usage{
prompt_default_sys(...)
}
\arguments{
\item{...}{Additional arguments passed to the template as data.}
}
\value{
\itemize{
\item \code{prompt_default_sys()}: A character string representing the default system prompt.
}
}
\description{
A collection of functions for generating prompts for the chat agent.

Functions:
\itemize{
\item \code{prompt_default_sys()}: Generate the default system prompt for the chat session.
}
}
\seealso{
\code{\link[ellmer:interpolate]{ellmer::interpolate_file()}}
}
