% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stores.R
\name{stores}
\alias{stores}
\alias{openai_list_vector_stores}
\alias{openai_create_vector_store}
\alias{openai_retreive_vector_store}
\title{Vector Stores}
\usage{
openai_list_vector_stores(api_key = get_api_key(name = "openai"))

openai_create_vector_store(name, api_key = get_api_key(name = "openai"))

openai_retreive_vector_store(id, api_key = get_api_key(name = "openai"))
}
\arguments{
\item{api_key}{A character string specifying the OpenAI API key.}

\item{name}{A character string specifying the service name.}

\item{id}{A character string specifying the vector store ID.}
}
\value{
\itemize{
\item \code{openai_list_vector_stores()}: A tibble of vector store information.
\item \code{openai_create_vector_store()}: A list containing the status, ID, and name of the created vector store.
\item \code{openai_retreive_vector_store()}: A list containing the vector store information.
}
}
\description{
These functions are helpers for working with OpenAI Vector Stores.
\itemize{
\item \code{openai_list_vector_stores()}: Retrieve a list of vector stores.
\item \code{openai_create_vector_store()}: Create a new vector store.
\item \code{openai_retreive_vector_store()}: Retrieve a vector store by ID.
}
}
