% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webr_packages.R
\name{suggest_package_alternatives}
\alias{suggest_package_alternatives}
\title{Suggest Package Alternatives}
\usage{
suggest_package_alternatives(package_name)
}
\arguments{
\item{package_name}{Character string of unavailable package name}
}
\value{
Character vector of alternative package suggestions
}
\description{
Suggests alternative packages available in webR for unavailable packages.
}
