% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tools}
\alias{tools}
\alias{tool_extract_code}
\alias{tool_execute_r_code}
\alias{tool_query_db}
\alias{tool_gmaps_geocode_address}
\alias{tool_gmaps_places_search}
\title{Custom Tools}
\usage{
tool_extract_code(cache = FALSE)

tool_execute_r_code()

tool_query_db(conn)

tool_gmaps_geocode_address(use_cache = TRUE, cache = NULL)

tool_gmaps_places_search(use_cache = TRUE, cache = NULL)
}
\arguments{
\item{cache}{Logical. If \code{TRUE}, the function will cache the results of the tool call using \code{\link[memoise:memoise]{memoise::memoise()}}.
Default is \code{FALSE}.}
}
\value{
Each function returns an \code{\link[ellmer:tool]{ellmer::tool()}} definition.
}
\description{
Functions that create custom \code{\link[ellmer:tool]{ellmer::tool()}} definitions.

Tool Functions:
\itemize{
\item \code{tool_extract_code()}: Tool that extracts code blocks from text. Wraps the \code{\link[=extract_code]{extract_code()}} function.
}
}
