% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{trim_code}
\alias{trim_code}
\title{Trim Code Block}
\usage{
trim_code(code)
}
\arguments{
\item{code}{The code block to trim.}
}
\value{
A trimmed code block.
}
\description{
Removes leading/trailing whitespace and empty lines from a code block.
}
