% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webr_packages.R
\name{webr_packages}
\alias{webr_packages}
\title{WebR Package Management Functions}
\description{
Functions for checking package availability and managing dependencies
in the webR/Shinylive environment.
}
