% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{xl_get_metadata}
\alias{xl_get_metadata}
\title{Get Excel Metadata}
\usage{
xl_get_metadata(xl_file, sheets = NULL)
}
\arguments{
\item{xl_file}{Path to an Excel workbook.}

\item{sheets}{A character vector of sheet names to include in the metadata. If \code{NULL}, all sheets will be included.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{file_path}: The normalized path to the Excel workbook.
\item \code{file_info}: A list of file information.
\item \code{xl_sheets}: A character vector of sheet names.
\item \code{xl_cells}: A data frame of cell information.
\item \code{xl_formats}: A data frame of cell formats.
\item \code{xl_named_ranges}: A data frame of named ranges.
}
}
\description{
Get metadata from an Excel workbook using the \link{tidyxl} package.
}
