% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checks}
\alias{checks}
\alias{check_inherits}
\alias{check_installed}
\alias{check_chat}
\alias{check_tool}
\alias{check_row}
\title{Check Functions}
\usage{
check_inherits(
  x,
  class,
  x_arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

check_installed(pkg, arg = rlang::caller_arg(pkg), call = rlang::caller_env())

check_chat(chat, arg = rlang::caller_arg(chat), call = rlang::caller_env())

check_tool(tool, arg = rlang::caller_arg(tool), call = rlang::caller_env())

check_row(row, arg = rlang::caller_arg(row), call = rlang::caller_env())
}
\arguments{
\item{x}{An object to check.}

\item{class}{A character string representing the class to check against.}

\item{x_arg}{A character string representing the argument name.}

\item{call}{An environment representing the calling environment.}

\item{pkg}{A character string representing the package name.}

\item{chat}{An \link[ellmer:Chat]{ellmer::Chat} object.}

\item{tool}{An \code{\link[ellmer:tool]{ellmer::tool()}} object.}

\item{row}{A single-row data frame.}
}
\description{
A collection of functions for checking the validity of function arguments.

Functions:
\itemize{
\item \code{check_inherits()}: Check if an object inherits from a class.
\item \code{check_installed()}: Check if a package is installed.
\item \code{check_chat()}: Check if an object is an \link[ellmer:Chat]{ellmer::Chat} object.
\item \code{check_tool()}: Check if an object is an \code{\link[ellmer:tool]{ellmer::tool()}} object.
\item \code{check_row()}: Check if an object is a single-row data frame.
}
}
