% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\alias{get_openai_api_key}
\alias{set_openai_api_key}
\alias{get_gemini_api_key}
\alias{set_gemini_api_key}
\alias{get_gmaps_api_key}
\alias{set_gmaps_api_key}
\title{Configuration}
\usage{
get_openai_api_key()

set_openai_api_key(key)

get_gemini_api_key()

set_gemini_api_key(key)

get_gmaps_api_key()

set_gmaps_api_key(key)
}
\value{
Each function returns the respective API key or sets it as an environment variable.
}
\description{
Functions for managing configuration settings and environment variables.

Functions:
\itemize{
\item \code{get_openai_api_key()}: Get the OpenAI API key from the configuration file.
\item \code{set_openai_api_key()}: Set the OpenAI API key as an environment variable.
\item \code{get_gemini_api_key()}: Get the Gemini API key from the configuration file.
\item \code{set_gemini_api_key()}: Set the Gemini API key as an environment variable.
\item \code{get_gmaps_api_key()}: Get the Google Maps API key from the configuration file.
\item \code{set_gmaps_api_key()}: Set the Google Maps API key as an environment variable.
\item \code{get_hunterio_api_key()}: Get the Hunter.io API key from the configuration file.
\item \code{set_hunterio_api_key()}: Set the Hunter.io API key as an environment variable.
}
}
