% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{npx}
\alias{npx}
\title{Run \code{npx} Command}
\usage{
npx(cmd, ..., .timeout = 60, .echo = TRUE)
}
\arguments{
\item{cmd}{The command to run.}

\item{...}{Additional arguments to pass to the command.}

\item{.timeout}{The timeout in seconds for the command to complete. Default is 60 seconds.}

\item{.echo}{Logical. If \code{TRUE}, the output will be printed to the console. Default is \code{TRUE}.}
}
\value{
The output of the command.
}
\description{
This function runs a command using \code{npx} and returns the output.
}
\examples{
\dontrun{
# get npx help
npx("--help")

# create a new react app
npx("create-react-app", "my-app")

# run repomix tool
npx("repomix", "--help")
}
}
