% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_sys}
\alias{pkg_sys}
\title{Package System File}
\usage{
pkg_sys(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{system.file()}.}
}
\value{
A character string representing the path to the system file.
}
\description{
Get the path to a system file within the package.
}
\examples{
pkg_sys("config/config.yml")
}
