% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tools}
\alias{tools}
\alias{tool_current_time}
\alias{tool_geocode_location}
\alias{tool_get_weather}
\alias{tool_execute_r_code}
\alias{tool_extract_code}
\alias{tool_query_db}
\alias{tool_gmaps_geocode_address}
\alias{tool_gmaps_places_search}
\alias{tool_anomaly_detection}
\alias{tool_correlate_with_anomalies}
\title{Custom Tools}
\usage{
tool_current_time(tz = NULL)

tool_geocode_location(
  use_cache = FALSE,
  cache = NULL,
  api_key = get_api_key("openweather")
)

tool_get_weather(
  use_cache = FALSE,
  cache = NULL,
  api_key = get_api_key("openweather")
)

tool_execute_r_code(cache = FALSE)

tool_extract_code(cache = FALSE)

tool_query_db(conn, cache = FALSE)

tool_gmaps_geocode_address(use_cache = TRUE, cache = NULL)

tool_gmaps_places_search(use_cache = TRUE, cache = NULL)

tool_anomaly_detection(data)

tool_correlate_with_anomalies(data)
}
\arguments{
\item{tz}{Timezone to use for the current time. Default is \code{NULL}, which uses the system timezone.}

\item{use_cache}{Logical. If \code{TRUE}, the function will cache the results of the tool call using \code{\link[memoise:memoise]{memoise::memoise()}}.
Default is \code{FALSE}.}

\item{cache}{Cache object for memoization. Default is \code{NULL}, which uses the default cache.}

\item{api_key}{API key for the OpenWeather API. Defauls to \code{get_api_key("openweather")}.}

\item{conn}{Database connection object. Default is \code{NULL}, which uses the default connection.}

\item{data}{Data frame to analyze for anomalies. Default is \code{NULL}.}
}
\value{
Each function returns an \code{\link[ellmer:tool]{ellmer::tool()}} definition.
}
\description{
Functions that create custom \code{\link[ellmer:tool]{ellmer::tool()}} definitions.

Tool Functions:
\itemize{
\item \code{tool_current_time()}: Tool that returns the current time.
\item \code{tool_get_weather()}: Tool that retrieves the current weather for a given location.
Wraps the \code{\link[=get_weather]{get_weather()}} function.
\item \code{tool_extract_code()}: Tool that extracts code blocks from text. Wraps the \code{\link[=extract_code]{extract_code()}} function.
\item \code{tool_execute_r_code()}: Tool that executes R code and returns the result.
\item \code{tool_query_db()}: Tool that executes a SQL query on the database and returns the result.
\item \code{tool_gmaps_geocode_address()}: Tool that geocodes an address using the Google Maps Geocoding API.
Wraps the \code{\link[=gmaps_geocode_address]{gmaps_geocode_address()}} function.
\item \code{tool_gmaps_places_search()}: Tool that searches for a place using the Google Maps Places API.
Wraps the \code{\link[=gmaps_places_search]{gmaps_places_search()}} function.
\item \code{tool_anomaly_detection()}: Tool that detects anomalies in a dataset.
\item \code{tool_correlate_with_anomalies()}: Tool that analyzes correlations and differences between anomalies and normal data.
}
}
\examples{
examples/ex_tools.R
}
\seealso{
\code{\link[ellmer:tool]{ellmer::tool()}}
}
